feat(refactor): replaced ObservablePipe and PromisePipe with AsyncPipe
