+++ modules/angular2/pipes.ts
@@ -4,10 +4,9 @@
  * This module provides advanced support for extending change detection.
  */
 
-export {PromisePipe} from './src/change_detection/pipes/promise_pipe';
 export {UpperCasePipe} from './src/change_detection/pipes/uppercase_pipe';
 export {LowerCasePipe} from './src/change_detection/pipes/lowercase_pipe';
-export {ObservablePipe} from './src/change_detection/pipes/observable_pipe';
+export {AsyncPipe} from './src/change_detection/pipes/async_pipe';
 export {JsonPipe} from './src/change_detection/pipes/json_pipe';
 export {DatePipe} from './src/change_detection/pipes/date_pipe';
 export {DecimalPipe, PercentPipe, CurrencyPipe} from './src/change_detection/pipes/number_pipe';
